First Commit
Basically I’ve broken the whole project but I was hoping to get Zxing
barcode installed. I have a lot of errors, I think it might be the
pom.xml file but I can’t bloody open it in the document folders or find
it in the solution itself…illusive bastard.